feat(cache): import/export feature to stabilize tests
