+++ .gitignore
@@ -27,7 +27,6 @@ heimdall/scripts/update
 heimdall/scripts/git
 
 coverage.xml
-*.bin
 *.o
 false
 false/*
