+++ bifrost/bifrost
@@ -25,13 +25,6 @@ main() {
                 ;;
             -v|--version)     shift; TARGET_VERSION=$1 ;;
             -B|--binary|--bin) shift; USE_BINARY=true ;;
-            -b|--bench)
-                benchmark
-                exit 0
-                ;;
-            -t|--test)
-                test
-                ;;
             -h|--help)
                 usage
                 exit 0
@@ -59,15 +52,14 @@ main() {
     # remove the source directory if it exists
     ensure rm -rf "$BUILD_PATH/heimdall-rs"
 
+    # clone heimdall-rs and cd into it
     cd $BUILD_PATH
     echo "bifrost: cloning 'Jon-Becker/heimdall-rs'."
     ensure git clone "https://github.com/Jon-Becker/heimdall-rs" > /dev/null 2>&1
-
     cd "heimdall-rs"
-
     ensure git fetch origin
 
-    # if they specified a version, use that
+    # if they specified a version, checkout that tag or branch
     if [ -n "$TARGET_VERSION" ]; then
         echo "bifrost: installing version $TARGET_VERSION."
 
@@ -83,11 +75,13 @@ main() {
         ensure git checkout $tag -b latest > /dev/null 2>&1
     fi
 
+    # if the user wants to use the precompiled binary, download it
     if [ -n "$USE_BINARY" ]; then
-
+        # cd into the binary directory
         ensure cd $BIFROST_BIN_DIR
         echo "bifrost: fetching binary."
 
+        # download the binary
         if [[ "$OSTYPE" == "linux-gnu"* ]]; then
             ensure curl -k -L -s --compressed "https://github.com/Jon-Becker/heimdall-rs/releases/download/$TARGET_VERSION/heimdall-linux-amd64" -o heimdall
         elif [[ "$OSTYPE" == "darwin"* ]]; then
@@ -96,7 +90,6 @@ main() {
             echo "bifrost: unsupported operating system: $OSTYPE"
             exit 1
         fi
-
         echo "bifrost: installing binary."
 
         # make the binary executable
@@ -113,6 +106,7 @@ main() {
         if [ $VERSION -ge 60 ]; then
             RUSTFLAGS="-C target-cpu=native -C codegen-units=1" CARGO_PROFILE_RELEASE_LTO=true ensure cargo install --path ./cli --bins --locked --force --root $BIFROST_PATH
         else
+            # using legacy build system
             echo "bifrost: installing with old build system."
             RUSTFLAGS="-C target-cpu=native -C codegen-units=1" CARGO_PROFILE_RELEASE_LTO=true ensure cargo install --path ./heimdall --locked --force --root $BIFROST_PATH
         fi
@@ -150,20 +144,21 @@ OPTIONS:
     -B, --binary    Install a precompiled binary instead of building from source
     -v, --version   Install a specific version
     -l, --list      List all available versions
-    -b, --bench     Install and run benchmarks
-    -t, --test      Install and run tests
 
 EOF
 }
 
+# ensure runs a command and exits if it fails
 ensure() {
     if ! "$@"; then echo "bifrost: required command '$*' failed."; exit 1; fi
 }
 
+# command_exists checks if a command exists
 command_exists() {
     command -v "$1" > /dev/null 2>&1
 }
 
+# requires_cmd checks if a command exists and exits if it doesn't
 requires_cmd() {
     if ! command_exists "$1"; then
         echo "bifrost: '$1' is required but not installed on this system"
@@ -171,99 +166,5 @@ requires_cmd() {
     fi
 }
 
-benchmark() {
-    echo "bifrost: fetching..."
-    # remove the current heimdall installation if it exists
-    ensure rm -f "$BIFROST_BIN_DIR/heimdall"
-
-    # make the build path if it doesn't exist
-    BUILD_PATH="${BIFROST_PATH}/build"
-    if [ ! -d $BUILD_PATH ]; then
-        ensure mkdir -p $BUILD_PATH
-    fi
-
-    # remove the source directory if it exists
-    ensure rm -rf "$BUILD_PATH/heimdall-rs"
-
-    cd $BUILD_PATH
-
-    ensure git clone "https://github.com/Jon-Becker/heimdall-rs" > /dev/null 2>&1
-
-    cd "heimdall-rs"
-
-    ensure git fetch origin > /dev/null 2>&1
-
-    # if they specified a version, use that
-    if [ -n "$TARGET_VERSION" ]; then
-        ensure git checkout $TARGET_VERSION > /dev/null 2>&1
-    else
-
-        # checkout the latest tag
-        tag=$(git describe --tags `git rev-list --tags --max-count=1`)
-        ensure git checkout $tag -b latest > /dev/null 2>&1
-    fi
-
-    echo "bifrost: compiling..."
-    ensure cd $BUILD_PATH/heimdall-rs/heimdall
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true ensure cargo build --release --package heimdall > /dev/null 2>&1
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true ensure cargo build --release --package heimdall-config  > /dev/null 2>&1
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true ensure cargo build --release --package heimdall-common  > /dev/null 2>&1
-    echo "bifrost: running benchmarks..."
-
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true cargo test --release --package heimdall -- benchmark_ | grep -E "±|benchmark_([a-zA-Z_])*:" >> stdout
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true cargo test --release --package heimdall-config -- benchmark_ | grep -E "±|benchmark_([a-zA-Z_])*:" >> stdout
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true cargo test --release --package heimdall-common -- benchmark_ | grep -E "±|benchmark_([a-zA-Z_])*:" >> stdout
-    clear
-    echo "bifrost: benchmark results:\n"
-    cat stdout
-    rm stdout
-    exit 0
-}
-
-test() {
-    echo "bifrost: fetching..."
-
-    # remove the current heimdall installation if it exists
-    ensure rm -f "$BIFROST_BIN_DIR/heimdall"
-
-    # make the build path if it doesn't exist
-    BUILD_PATH="${BIFROST_PATH}/build"
-    if [ ! -d $BUILD_PATH ]; then
-        ensure mkdir -p $BUILD_PATH
-    fi
-
-    # remove the source directory if it exists
-    ensure rm -rf "$BUILD_PATH/heimdall-rs"
-
-    cd $BUILD_PATH
-
-    ensure git clone "https://github.com/Jon-Becker/heimdall-rs" > /dev/null 2>&1
-
-    cd "heimdall-rs"
-
-    ensure git fetch origin > /dev/null 2>&1
-
-    # if they specified a version, use that
-    if [ -n "$TARGET_VERSION" ]; then
-        ensure git checkout $TARGET_VERSION > /dev/null 2>&1
-    else
-
-        # checkout the latest tag
-        tag=$(git describe --tags `git rev-list --tags --max-count=1`)
-        ensure git checkout $tag -b latest > /dev/null 2>&1
-    fi
-
-    echo "bifrost: compiling..."
-    ensure cd $BUILD_PATH/heimdall-rs/heimdall
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true ensure cargo build --release --package heimdall > /dev/null 2>&1
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true ensure cargo build --release --package heimdall-config  > /dev/null 2>&1
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true ensure cargo build --release --package heimdall-common  > /dev/null 2>&1
-    echo "bifrost: running tests..."
-
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true cargo test --release --package heimdall -- test_ --nocapture
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true cargo test --release --package heimdall-config -- test_ --nocapture
-    RUSTFLAGS="-C target-cpu=native" CARGO_PROFILE_RELEASE_LTO=true cargo test --release --package heimdall-common -- test_ --nocapture
-    exit 0
-}
-
+# run main
 main "$@" || exit 1
